Use std::sin/cos as shown in 9fcd7f2dc05fd63a92c4e036ddbb866ec6ec5d72

+++ src/player.cpp
@@ -285,5 +285,5 @@ bool PlayerPhysicsOverride::operator==(const PlayerPhysicsOverride &other) const
 
 v2f PlayerControl::getMovement() const
 {
-	return v2f(sinf(movement_direction), cosf(movement_direction)) * movement_speed;
+	return v2f(std::sin(movement_direction), std::cos(movement_direction)) * movement_speed;
 }
